+++ lubee/src/onboarding/birth/index.tsx
@@ -1,11 +1,11 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import OnboardingHeader from "../components/OnboardingHeader";
 import ProgressBar from "../components/ProgressBar";
 import OnboardingTitleBox from "../components/OnboardingTitleBox";
-import YellowBox from "../components/YellowBox";
 import OnboardingBtn from "../components/OnboardingBtn";
+import DateInput from "../components/DateInput";
 
 interface BirthProps {
   moveToOnboardingProfile: () => void;
@@ -21,34 +21,6 @@ export default function index(props: BirthProps) {
 
   const isOnboardingBtnDisabled = !(year && month && day);
 
-  // 오늘보다 미래 선택 시 오늘 날짜로 변경
-  useEffect(() => {
-    if (year.length === 4 && month.length === 2 && day.length === 2) {
-      const inputDate = new Date(`${year}-${month}-${day}`);
-      const today = new Date();
-
-      if (inputDate > today) {
-        const todayStr = today.toISOString().split("T")[0].split("-");
-        setYear(todayStr[0]);
-        setMonth(todayStr[1]);
-        setDay(todayStr[2]);
-      }
-    }
-  }, [year, month, day]);
-
-  // 한자리 수 입력 시 0 붙이기
-  useEffect(() => {
-    if (month.length === 1 && month !== "0") {
-      setMonth(`0${month}`);
-    }
-  }, [month]);
-
-  useEffect(() => {
-    if (day.length === 1 && day !== "0") {
-      setDay(`0${day}`);
-    }
-  }, [day]);
-
   function handleBackBtn() {
     moveToOnboardingProfile();
   }
@@ -61,20 +33,6 @@ export default function index(props: BirthProps) {
     moveToOnboardingAnniv();
   }
 
-  // 월은 01~12로
-  function handleMonthChange(value: string) {
-    if (value === "" || (value.length <= 2 && (parseInt(value) >= 1 || value === "0"))) {
-      setMonth(value);
-    }
-  }
-
-  // 일은 01~31로
-  function handleDayChange(value: string) {
-    if (value === "" || (value.length <= 2 && (parseInt(value) >= 1 || value === "0"))) {
-      setDay(value);
-    }
-  }
-
   return (
     <Wrapper>
       <OnboardingHeader handleBackBtn={handleBackBtn} handleXBtn={handleXBtn} showBackIcon showXIcon />
@@ -83,11 +41,7 @@ export default function index(props: BirthProps) {
         titleText="본인의 생년월일을 입력해주세요"
         subtitleText="달력에 나와 연인의 생일이 표시돼요"
       />
-      <YellowBoxContainer>
-        <YellowBox inputValue={year} setInputValue={setYear} $disabled={true} placeholder="YYYY" />
-        <YellowBox inputValue={month} setInputValue={handleMonthChange} $disabled={true} placeholder="MM" />
-        <YellowBox inputValue={day} setInputValue={handleDayChange} $disabled={true} placeholder="DD" />
-      </YellowBoxContainer>
+      <DateInput year={year} setYear={setYear} month={month} setMonth={setMonth} day={day} setDay={setDay} />
       <OnboardingBtn handleOnboardingBtn={handleOnboardingBtn} text="다음" $disabled={isOnboardingBtnDisabled} />
     </Wrapper>
   );
@@ -100,10 +54,4 @@ const Wrapper = styled.div`
   position: relative;
   width: 100%;
   height: 100%;
-`;
-
-const YellowBoxContainer = styled.section`
-  display: flex;
-  gap: 1.2rem;
-  margin-top: 6rem;
 `;
